config: move config-related code from utils
no reason to splatter the code for the configuration handling among
different packages, try to regroup everything under config.
