+++ subcommands/config/config.go
@@ -30,8 +30,8 @@ import (
 	"github.com/PlakarKorp/kloset/connectors/importer"
 	"github.com/PlakarKorp/kloset/connectors/storage"
 	"github.com/PlakarKorp/plakar/appcontext"
+	"github.com/PlakarKorp/plakar/config"
 	"github.com/PlakarKorp/plakar/subcommands"
-	"github.com/PlakarKorp/plakar/utils"
 	"go.yaml.in/yaml/v3"
 	"gopkg.in/ini.v1"
 )
@@ -112,7 +112,7 @@ func dispatchSubcommand(ctx *appcontext.AppContext, cmd string, subcmd string, a
 			}
 			cfgMap[name][key] = val
 		}
-		return utils.SaveConfig(ctx.ConfigDir, ctx.Config)
+		return config.Save(ctx.ConfigDir, ctx.Config)
 
 	case "check":
 		p := flag.NewFlagSet("check", flag.ExitOnError)
@@ -201,7 +201,7 @@ func dispatchSubcommand(ctx *appcontext.AppContext, cmd string, subcmd string, a
 			thirdParty = "rclone"
 		}
 
-		newConfMap, err := utils.GetConf(rd, thirdParty)
+		newConfMap, err := config.LoadFile(rd, thirdParty)
 		if err != nil {
 			return fmt.Errorf("failed to load config: %w", err)
 		}
@@ -242,7 +242,7 @@ func dispatchSubcommand(ctx *appcontext.AppContext, cmd string, subcmd string, a
 				}
 			}
 		}
-		return utils.SaveConfig(ctx.ConfigDir, ctx.Config)
+		return config.Save(ctx.ConfigDir, ctx.Config)
 
 	case "ping":
 		p := flag.NewFlagSet("ping", flag.ExitOnError)
@@ -323,7 +323,7 @@ func dispatchSubcommand(ctx *appcontext.AppContext, cmd string, subcmd string, a
 			return fmt.Errorf("%s %q does not exist", cmd, name)
 		}
 		delete(cfgMap, name)
-		return utils.SaveConfig(ctx.ConfigDir, ctx.Config)
+		return config.Save(ctx.ConfigDir, ctx.Config)
 
 	case "set":
 		p := flag.NewFlagSet("set", flag.ExitOnError)
@@ -348,7 +348,7 @@ func dispatchSubcommand(ctx *appcontext.AppContext, cmd string, subcmd string, a
 			}
 			cfgMap[name][key] = val
 		}
-		return utils.SaveConfig(ctx.ConfigDir, ctx.Config)
+		return config.Save(ctx.ConfigDir, ctx.Config)
 
 	case "show":
 		var opt_json bool
@@ -452,7 +452,7 @@ func dispatchSubcommand(ctx *appcontext.AppContext, cmd string, subcmd string, a
 			}
 			delete(cfgMap[name], key)
 		}
-		return utils.SaveConfig(ctx.ConfigDir, ctx.Config)
+		return config.Save(ctx.ConfigDir, ctx.Config)
 
 	default:
 		return fmt.Errorf("usage: plakar %s [add|check|import|ping|rm|set|show|unset]", cmd)
